Test: 요금 계산 통합테스트 -> 단위테스트로 변경
- 코드 리뷰 반영: 단위테스트로 변경
